Remove dead code, rename the print button component to PrintButton
