+++ client/src/components/ProjectWizard/WizardFooter.js
@@ -3,7 +3,7 @@ import PropTypes from "prop-types";
 import NavButton from "../Button/NavButton";
 import SaveButton from "../Button/SaveButton";
 import { createUseStyles } from "react-jss";
-import DownloadButton from "../Button/DownloadButton";
+import PrintButton from "../Button/PrintButton";
 import ReactToPrint from "react-to-print";
 import { PdfPrint } from "../PdfPrint/PdfPrint";
 
@@ -38,7 +38,7 @@ const WizardFooter = ({
   setDisabledForNextNavButton,
   setDisabledSaveButton,
   setDisplaySaveButton,
-  setDisplayDownloadButton,
+  setDisplayPrintButton,
   onSave,
   dateModified
 }) => {
@@ -77,9 +77,9 @@ const WizardFooter = ({
             </div>
             <ReactToPrint
               trigger={() => (
-                <DownloadButton
-                  id="downloadButton"
-                  isDisplayed={setDisplayDownloadButton()}
+                <PrintButton
+                  id="PrintButton"
+                  isDisplayed={setDisplayPrintButton()}
                 />
               )}
               content={() => componentRef.current}
@@ -115,7 +115,7 @@ WizardFooter.propTypes = {
   setDisabledForNextNavButton: PropTypes.any,
   setDisabledSaveButton: PropTypes.any,
   setDisplaySaveButton: PropTypes.any,
-  setDisplayDownloadButton: PropTypes.any,
+  setDisplayPrintButton: PropTypes.any,
   onSave: PropTypes.any,
   onDownload: PropTypes.any,
   dateModified: PropTypes.any
